requirements.txt
numpy
openai
pandas
python-dotenv
